base: add integer conversion utilites for StringView objects
Similar in motivation to aosp/3391179, it is not currently possible
to use the existing string utility functions to convert StringViews
to integers without creating a copy of the string the in process.

To rectify this, add a new set of utility functions to convert
StringView objects to integers. These functions are tested using
identical unit tests as the existing std::string versions of the
functions.

A conversion routine for converting StringViews to doubles is
deliberately omitted due to a limitation of clang's standard lib.
Code comments are left in place explaining its absence. When clang
gains full C++ 17 support, this double conversion routine can be
safely enabled.

Change-Id: Ic89bb192da72c0a67291eb7855ae06ff0be9885a