tp: fix parsing battery stats history strings that contain commas
Fix parsing of battery stats history strings that cointain commas.
Previously, history strings that contained commas would
unintentionally tokenized by the StringSplitter.

to fix this we process the last element of HSP lines manually. To
facilite the ability to parse the remainder of the line manually, a
new remainder() API is added to string splitter to aid parsing.

As part of the fix, we are now adopting the convention that strings
stored by the HistoryStringTracker will have their quotes removed
first. This makes more sense as eventually not every caller will
want the strings quoted.

Change-Id: Ifb5fb565ebfd205eef6575d10b749e36635b984c