tp: import wakelocks from battery stats
Add wakelocks from battery stats in a similar manner to how they
are emitted from atrace, using the same "WakeLocks" name for the
track. The main difference is here we are using a global track
instead of a process track for the process that emitted the
wakelock. This is partially becuase we do not have PIDs available
from battery stats, so we can't create process tracks.

Wakelocks parsing is only available on battery stats ver 36 and
above due to us having no "cookie" to match begin & end async
slices with on prior versions.

Change-Id: Ib56f5f97d509b18a5e7aba1bb8c8c3574a23f7bc