ui: Convert track event to area selection when marking a span with 'M'
A common user workflow is to hit 'M' to turn the current selection into
a marked area and then use the checkboxes to add more tracks to the
area, in order to see the aggregated stats in the details panel.

This workflow was broken when we removed track lists from marked areas,
effectively turning marked areas into simple time spans.

Presumably users don't actually want to create a marked area, they just
want to convert a slice selection into an area selection. Creating a
mark was just the quickest way to do this. However, the muscle memory
persists.

This CL patches the issue by making 'M' also create an area selection
as well as creating the marked area.

Bug: https://github.com/google/perfetto/issues/884
Change-Id: I8f60f8b437dcc565caac3ae29474b4959cfe5f14