tp: parse battery stats ints as signed ints
The battery stats checkin is being emitted and serialized in Java.
Java does not support unsigned integers and thus all integers in
the Battery stats checkin should be signed.

Parse these integers as signed integers becase it is safer to do so
in the case the one of these fields somehow winds up negative.

Change-Id: I73e1a8b9418af355049c2f445982bae439eff4d2