move battery stats event count into different diff test
Currently we have one bugreport test that validates battery stats,
however if there is any change in the number of events emitted which
is not expected, then it can be come difficult to tell which events
were added or removed by looking at the diff since the count is
included on every line, causing the diff to flag every line as
changed.

In order to improve the readability of diff test output upon
failure, move the event count validation to its own test case.

Change-Id: I6456f7305b2d1850022a7fc6cec45fc356b10d56