base: handle negative number strings when converting to unsigned
StringView conversion to unsigned integers from signed strings
(eg. -12345) will fail, but this will succeed with the existing
string conversion routines.

Update the string view logic to make them also succeed in this case
so that we have parity between the various conversion
implementations.

Also add unit tests that validate the behavior of a negative number
input to the unsigned integer conversion functions.

Change-Id: Id1b83f1d0b45624d7109a80c97095b5c44d0d577