ui: Create SimpleColumn
The goal of creating SimpleColumn is to have an alternative to `well_known_columns.ts` columns, which are very specialised implementations of the column and should live in `org.Chromium.OpenTable`.

Also marked components/table as `components/legacy_table` to have a clean distinction between new and old implementations. At the end of this refactor, all files from there should be either removed or moved into aforementioned plugin.

Change-Id: I35ae848dbfb6e302805169b2feecadebc63a7711