tp: share ArgsTracker/GlobalArgsTracker in multi-machine traces
The `arg_set_id` column in the `args` table is expected to be
monotonically increasing. This property was not maintained when loading
multi-machine traces, leading to incorrect (decreasing) `arg_set_id`
values.

This commit ensures that multiple machines share the same
ArgsTracker/GlobalArgsTracker objects, preventing duplicate and
out-of-order `arg_set_id` assignments.

This fixes a bug where the `cpu_counter_track` table, which relies on
`dimensions_arg_set_id`, failed to load correctly for multi-machine
traces.

Bug: 383043300
Bug: 284258446

Change-Id: I0efc3320b312fbafe9b2e1c160e60c1529364a1f