tracing_service_impl: Service event when flush for clone is done
With this, TracingService emits a service event when the flush related
to a clone finishes. The event is emitted only on the cloned tracing
session.

This is useful to understand how long cloning takes.

Since different buffers can finish a clone at different times, the event
is emitted once per buffer.

The change is not as straightforward because, when we start the flush,
the cloned tracing session doesn't exist yet. So the timestamp has to be
plumbed through a couple of places.

Bug: 383733250
Change-Id: Ic063c000819045a8c61a9948e6214de20591f5ee