tracing_service_impl: Service event when flush for clone starts
With this, TracingService emits a service event when the flush related
to a clone starts. The event is emitted only on the cloned tracing
sessions.

This is useful to understand how long cloning takes.

The newly introduced `TracePacket.clone_snapshot_trigger` reports a
similar timestamp, but, in case of non zero `Trigger.stop_delay_ms`,
it's not exactly what we want.

The change is not as straightforward because, when we start the flush,
the cloned tracing session doesn't exist yet. So the timestamp has to be
plumbed through a couple of places.

Bug: 383733250
Change-Id: Id364c4805d75d744fb1636ca0f356e71de8825bb