ui: Add missing ScreenState enums to battery plugin
Looking at AOSP source code it appears the ScreenState track is
getting populated by values coming from enums in Display.java there
are a few enums that Perfetto does not handle which sometimes
results in an "Unknown" screenstate.

https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/android/view/Display.java;l=443-504;drc=493a723df2a3affb653732e0158ad1ac32e10931

Fill in all the remaining enums listed in Display.java to make it
less likely that users see "Unknown" on the screen state track.

Change-Id: I35e08a44bff5d3b13f5e0ef241d2fe0dbc80bbbc