ui: Fix crash during area selection on hidden customSqlTableSliceTracks
`CustomSqlTableSliceTrack`s create an ephemeral view in their
`onCreate()` hook which is used for rendering. This view is created when
the track first becomes visible on the timeline and dropped when it is
no longer visible. The view only exists to facilitate automatic depth
calculation for proper rendering - it is unnecessary for aggregations.

Currently, the exposed dataset from this track relies on this ephemeral
view, so making area selections on this track when it's hidden (e.g.
when selecting a collapsed group that contains one of these tracks)
causes a crash.

This CL changes the exposed dataset to be based on the source table for
this track (provided by `getSqlDataSource()`) rather than the view,
which should always exist.

Changes:

1. Change return type of `getSqlDataSource()` so it can no longer return
   a promise (no implementations relied on this anyway). This function
   needs to be synchronous because `getDataset()` is synchronous and now
   needs to call it.
2. Remove the `getSqlImports()` function. It was only being used in one
   place and it would mean that the necessary SQL includes might be
   missing that are required by the dataset. In the single instance that
   did use this feature, I just moved the include to the plugin that
   creates the track in the first place.
3. Expose the new dataset based on `getSqlDataSource()`.

Bug: 380480657
Change-Id: I15c22e910b0e78c7b37cdf9cd9978d0304aec144