tp: migrate legacy async tracks to new system and delete legacy code
This CL migrates the last remaining track type to the new tracks system
and the deletes all the legacy code from the codebase. This marks the
final CL of track migration of all code in trace processor.

Follow up CLs will rename classification -> type and make the UI
properly use this new system.

Change-Id: Ic1ac62eeac2c94aece92273ff2e2106b36917d63