Remove upid check for jank CUJ frame timeline table.
In certain cases, for eg. the new Desktop Windowing CUJs, it is observed
that the Surface(layer) that is being associated with the CUJ, is
generated in a different process(system_server) than that of the CUJ
process(com.android.systemui). This causes the jank timeline_metrics to
not be populated.

To fix this issue, the upid check between the actual frame slice and the
jank cuj vsync boundary table has been removed. The frame timeline table
will be populated by comparing the layer id between the above two
tables, and with vsync range check.

Test: manual test on a trace
Bug: 381499490
Change-Id: If95196201642daa295baf1e258dedb891fa0593b