ui: Fix track heights
Currently the track height is initialized before descriptor is in
TrackView's ctor, leading to all tracks having the default height.

To fix the issue and ensure it doesn't happen again, this CL moves the
getTrackHeight function outside the class, so that it doesn't depend on
the class properties at all. Generally avoiding calling member methods
within the ctor should help as this can hide out of order accesses
like this.

Change-Id: Iadc6c7ede33c0701d4de818acf1235c4ecc45af1