ui: remove obsolete recordingV2/ from v1 plugin
The RecordTrace plugin used to have two code-paths (one for v1 one
for v2) and a flag to internally switch over between the two.
This CL deletes most of the RecordTrace/recordingV2/ code and
repurposes the RecordingV2 flag to instead enable the new
RecordTraceV2 plugin.
This allows both plugins to be technically enabled by default, but
only one of them effectively takes action based on the flag.
This is so users that used to be on "old" recording v2 immediately get
the benefits of the "new v2" plugin.

Bug: 377651301
Change-Id: Id400255b6e4419a79dd03caf94ea39ad54727004