ui: Reduce LOD for offscreen tracks
This is a very simple implementation of virtual scrolling, where tracks
that lie outside the scrolling viewport render a simpler version of
themselves with a greatly reduced number of DOM elements and simpler
mithril view() function.

I.e. only the title is displayed, no buttons, menus,
chips, or ellipsised titles.

This means Ctrl+F to find tracks & scrolling to tracks still works just
as it used to because the track DOM nodes are still there.

During testing this reduces DOM render times by a factor of ~4x on
traces with a large number of tracks.

This feature is initially hidden behind a flag and disabled by default
to allow iteration.

Change-Id: I294fcf39e4b409ffc9be15e670b23fe8f98e22f4