tp: remove notion of kInvalidArgSetId
This CL removes the reservation of zero for invalid arg set ids. This
idea was designed for performance (as args columns then don't have to be
nullable) but with modern trace processor design, this prevents an
optimization which would significantly *improve* peformance.

Just move to a world where std::nullopt indcates no args instead.

Change-Id: I0bb2fbc6d7dd4564cb072dedee7e9c7146f6f82c