tp: rewrite graphics frame event parsing and remove frame_slice table
This CL rewrites all the graphics_frame_event parsing to be far far
easier to understand. The previous code was a complete mess and there
was no salvaging it without basically rewriting it. Thankfully there is
a pretty good test so we can be reasonably confident that there is no
obvious behaviour change as a result of this CL.

The reason for doing this was for removing the intrinsic frame_slice
table which is blocking the removal of parent/child tables in trace
processor. A rewrite was required to make this possible.

Change-Id: Ie640ad743777c4c51653bba02dd743b61dc4e1ab