ui: optimize single counter value selection queries
Using subqueries to extract data for the prev/next item in a counter
series significatly out performs LEAD/LAG in this case by roughly
one order of magnitude. Update these queries to use this method so
that they perform better.

By optimizing these queries, we make it less likely for a user
to accidentally trigger the race condition outlined in
https://github.com/google/perfetto/issues/981 athough the race
condidition is still present.

Change-Id: I07c537823f3a3bd729adf801f878605966e8c902