ui: Store track selection synchronously without waiting for details
This CL addresses two issues:

1. Avoid races when making selections:
- User selects a track event.
- Details start loading.
- User makes a different selection that loads quickly (i.e. area).
- Some time later the details finish loading and overwrite the latter
  selection.

2. Fixes the slight delay between clicking a track event and the
event being highlighted, which can be disconcerting and make the UI
feel unresponsive.

Note: This change makes the TrackEvent.details optional, as they can
take some time to load. Thus, any areas that use these details must
now deal with the fact that they can be possibly be undefined, but in
all cases we can do something sensible, so it's not a huge issue.

Bug: https://github.com/google/perfetto/issues/981
Change-Id: I04a051e4df3e1fe3264fed543c31dfa768156697