ui: support long responses from QueryServiceState
QueryServiceState is a RPC streaming response. I forgot about
that. This is usually not an issue when the response is <128KB
because, at the wire protocol level, there is no difference
between a one-shot response and a streaming-response with one chunk.
However, Eric hit cases where the QSS is > 128KB.

Bug: 389023563
Change-Id: I176d4e6a08b436ecc0f0ef6f12db82f163cc2e59