shared_lib: Fix emitting data with PerfettoDsOnStopArgsPostpone()
With the C data source API, the use user can register a callback when a
data source is stopped. In the callback, the user can postpone the stop
operation, by calling PerfettoDsOnStopArgsPostpone(step 1). Later, when
the user calls PerfettoDsStopDone(step 2), the stop operation is
considered complete.

Between step 1 and step 2, the user is supposed to be able to emit
more data. The implementation had a bug though: it always immediately
marked all the instances as disabled when receiving a stop signal.

Fix the problem by marking instances as disabled later, when the user
completes the stop operation.

This is the same behavior of the C++ data source API (covered by the
PerfettoApiTest.WriteEventsAfterDeferredStop integration test).

Reported-by: keanmariotti@google.com
Change-Id: I270a03d2ad10e8c095f9c445434ae02adfea3edb