tp: Parse dumpstate logcat before persistent logcat
We are planning to update the logcat readers to avoid sending
events to the sorter until after the timezone offset is know.
see:
https://github.com/google/perfetto/issues/934#issuecomment-2543441231

Because the information to compute the TZ offset is in dumpstate,
We opt to process that first to avoid needing to buffer all of
persistant logcat events.

For that reason this CL swaps the parse order of the parsing of
dumpstate and persistent logcat so to that we'll always parse
dumpstate first. This change is broken out separately from other
TZ related CLs to ensure that reording of the parsing logic does
not have any impact on  events that are output.

Change-Id: I2a578d3bcfc48b0a609c91f12eb2ca28bec5b8ac