tp: Optionally wait for TZ info before sending logs to sorter
Logcat events by default are not adjusted according to timezone,
and so if logcat events are sent to the before the trace timezone
offset is populated, then we'd assume a tz offset of zero, which
is wrong most of the time.

To help mitigate this issue we add new option that allows the logcat
reader to hold off on sending events to the sorter until either
the tz offset is populated, or we hit the end of input.

Change-Id: Id40671d11e778921b9df456befee0a6b3f67223f