tracing: clean up FD-passing when starting the service
Today we have a lot of plumbing to deal with the case of
Android Init passing the sockets as FDs. This caused the Start
method to have:
1. A variant that takes two strings (producer_name, consumer_name)
2. A variant that takes two fds (producer_fd, consumer_fd)
3. A variant that takes a vector of string, for the CrOS case that
   wants to listen on >1 producer socket.
Now we have a further use case coming from AAOS that wants to pass
a FD (for unix) + a string (for vsock) for the producer. This is getting
too complicated.
This CL simplifies the code using strings everywhere and using the

fd://NN convention to detal with the special case of android init.
Change-Id: Ie47a4de03de3c436d535a28487221f6bf21c79b8