ui: BaseSliceTrack - change rowSpec from a method to a constructor arg
The rowSpec is the object passed to QueryResult.iter() to define the
column names and types of the data within a query result's row.

Normally track implementations override BaseSliceTrack::getRowSpec(),
returning the rowSpec, but this can just as easily be passed in as a
constructor argument instead.

There are a couple of reasons why this makes sense:

1. No tracks ever need to change the rowSpec dynamically.
2. In upcoming CLs I need to be able to define the rowSpec from some
constructor args, which causes an issue as BaseSliceTrack calls
getRowSpec() from within it's own constructor, so any state this
function uses must be available when super() is called in the derived
track. However, javascript prohibits this to be used before super() is
called, leading to a chicken and egg situation.

Change-Id: Ia11446b9235ae863c5ea3cc7131dde962d832a89