ui: Add types to dataset, and add colorizer option to DatasetSliceTrack
This CL makes SourceDataset's typed by their schema. This allows the
dataset to carry information about its schema type, so that this type
can be used at compile time to infer other types and check that the
dataset is being used correctly.

The dataset type is normally automatically inferred by TS from its
schema, so callsites don't need to change.

This CL also updates DatasetSliceTrack so that any datatset passed in
will be checked at compile time whether it implements the correct
schema, which can catch bugs earlier and make the track easier to use.

This CL also adds an optional colorizer field in the track attrs, which
can be used to customize the color of each slice based on the value of
each row pulled out of TP. The type information comes in handy here as
the row which is passed to the colorizer can be typed using the
dataset's schema.

Change-Id: I96104feba066f80ac958754c6c09420125dfb092