ui: Add rootTableName property to tracks, move out of selectionDetails
TrackEventDetails currently contains a tableName which is a hack purely
to serve the flow manager.

In an effort to move towards a world where selection details are based
purely on the track's dataset, this CL moves this rootTableData to
the track itself (rather than relying on the track filling it in in the
selection details).

Having the rootTableName on the track itself will also serve to help
get the SQL selection resolvers working with dataset, which will be
done in an upcoming CL.

Change-Id: I7f8edc28ca34930aa77439227ad0d976f0f645c3