Provide an optional guest producer socket in perfetto Android builds
When Android is considered the host in a virtualized environment it
should be able to expose a producer socket for perfetto producers
running in guest machines. The port to be used can be specified by
setting the persist.traced.guest_producer_port system property to the
VSOCK or TCP address.

In a device's make file this can be specified as:
PRODUCT_PRODUCT_PROPERTIES += \
    persist.traced.guest_producer_port="vsock://-1:30001"

Test: build traced in Android tree and connect a producer in guest
      machine
Bug: 389792832
Change-Id: Ief245f053c28a0acfef10842c31169290e27e768