ui: Fix hovered slices not updating on mouse move
Tracks based on base_slice_track have a bug where hovering slices no
longer highlights the hovered slice (or slices with similar names in
some cases). This is due to the fact that the highlighted slice is only
updated on DOM redraws not canvas redraws.

We should fix this properly, but for now this CL just patches the issue
by kicking off a full redraw every mouse move in these tracks.

Change-Id: Ib76912c991ebe633cb37fa52d53ceb03f153070a