tp: integrate trace summarization into trace_processor_shell and diff tests
This CL does the necessary work to integrate the new summarization
system into trace_processor_shell and the diff testing frameworking.
While here, I had to make some non-trivial changes to the framework as
well to make it possible to do this cleanly.

I could hypothetically split these changes out but it's a lot of
trouble for not much gain given they are quite self contained and
it's just testing infra code.

Change-Id: I0cfa02c8ec7258b1079528497b9b59ba099f9214