implicit_segment_forest: builtins for MSVC
This uses GCC/clang builtins that are not available for MSVC.

We can use instead C++20 standard library (if available), or MSVC
builtins.

I would also provide a standard generic fallback, but it's a lot of
subtle code and when C++20 is going to be available everywhere, this
problem is going to go away.

Change-Id: Ica3f3b0e2686c9fcd9e9cf7abfe2dda1d522a4c5