protozero: Fix INT_MIN enum values with MSVC
Enum values such as `-2147483648` might be misinterpreted by some
compilers.

`-` might be interpreted as a unary minus and `2147483648` as an int
constant, but `2147483648` is too big to be a signed 32-bit integer.

Let's use the same trick as protobuf and generate the constant as
`-2147483647 -1`. See

https://github.com/protocolbuffers/protobuf/blob/336d6f04e94efebcefb5574d0c8d487bcb0d187e/src/google/protobuf/compiler/cpp/helpers.cc#L885

This fixes MSVC warning C4146.

Change-Id: I3d68a68f56552e0afdd3dd3647c1960fbaf361f9