ui: Port 'TraceProcessor' and 'ThreadSlice' tracks to DatasetSliceTrack
These tracks exercise the `fillRatio` feature of DatasetSliceTrack.

This CL also changes how the layout is provided to DatasetSliceTrack.
Instead of an enum, we now use a function that, given the dataset, returns a query that wraps the dataset and ensures a depth column is
present. For convenience, the following pre-canned depth provider
functions are provided:

- flatDepthProvider - provides depth columns whose value is always 0.
- internalLayoutDepthProvider - automatically calculates depth using
  internal_layout().

This allows a depth to be provided using `experimental_slice_layout`
which avoids risky changes to TraceProcessorSliceTrack.

This CL also removes the last usages of NamedSliceTrack, so this has
been removed.

Change-Id: Id42224475ff8b0f82a10573b406dae9864b037d2