traced_perf: relax CHECK on perf ring buffer record denseness
As of some kernel >6.6 (exact commit TBD), sampling on tracepoints
produces records with some zero padding at the end. This is correctly
included in the kernel's record header, but is causing traced_perf to
crash on a CHECK since we're verifying that the records are always as
dense as possible.

This patch simply relaxes the CHECK since it's proven to be too strict.
The ring buffer record reader (ReadUntilSample) already does the right
thing by using the kernel-supplied record size when advancing to the
next entry.

Bug: 376630540
Bug: 388539594
Change-Id: Ia4b61c526a5822204a625cf26eb47c3f561ee435