ui: Remove sqlEventResolvers, replace with dataset & rootTableName
This CL replaces the somewhat flakey sqlEventResolvers with an approach
based on using the track's dataset & rootTableName to work out which
track a given slice id on a given root able lives, and selecting it.

This approach moves the complexity into the core, rather than forcing
the plugins to deal with it which is a lot more reliable.

Fixes: 390653287
Change-Id: I8afa56b5544a9b0af7eeb4ccd2b9e72283fc900b