ui: Tidy up focusing on instant or DNF slices
When focusing on an instant event or DNF slice, we currently pretend
it's duration is 1ns or 30,000ns respectively which isn't useful to
anyone.

This CL cleans this up so when we focus on one of these events we just
pan to reveal the slice, never zoom, which means we don't have to make
any assumptions about how wide it is.

Change-Id: Ib53d36c09cc282f1d4950fc0f422246b265e0918