[tp] mipmap: throw an error if id doesn't fit into 32 bits
Currently the track just doesn't display anything if the id doesn't
fit into 32-bits, which is surprising.

Per conversation with Lalit, we don't want to support 64-bit ids in
trace processor for performance (and consistency) reasons, so just throw
an error, so it's clear to the user what is going on.

(For the immediate problem with the chrome scroll jank plugin,
aosp/3464774 fixes the underlying issue in the plugin).

Bug: 373819358
Change-Id: I476141243cddc1bdf512b4adc5f986ddef8771c6