ui: recording: don't crash if flush times out
Today when a Flush() request times out (e.g. because a process is
frozen or slow to respond), it NACKs the corresponding RPC request
in consumer_port.proto.
The UI throws an exception if any method fails, which is the right
thing to do in most other cases (we can't recover from an error in
EnableTracing/DisableTracing) but not here.
Add some code to allow Flush to fail silently.

Bug: 392075824
Change-Id: Ic374d534be749b0f90cca4d0ed40fb563bd1c960