ui: Move track menu items into a component to improve performance
Track menus typically need to do do an non insignificant amount of work
in to generate the necessary vnodes for the menu. Because of
the way popups work, these menu items must be rendered every frame for
every track regardless of whether the popup menu is actually open or
not.

This CL moves the menu items in a component, which takes advantage of
how mithril works and only calls the view function when the menu is
actually open. This makes a measurable difference when there are lots
of tracks on screen, but moreover allows us to add to this menu without
the risk of tanking performance.

Change-Id: I5fb262cbc6ca0b515fd5a6c96a40fa2dd6bce987