ui: Move rounding operation to very end
Move the final rounding before being displayed for the UI to the very
end, which will be more robust if things are added or modified in the
middle of the callflow in the future. Additionally, this will help with
the precision of the "Total Energy" column, since with this patch it
removes the redundant rounding done on idle cost that is later added to
create the total energy column.

Bug: 393391273
Change-Id: I824cbb203caace21d0821d8cbfd362b43745c007
Signed-off-by: Samuel Wu <wusamuel@google.com>