ui: Fix total energy double counting bug
The idle transition cost of the swapper thread was being counted twice
in the math for the total energy column.

This patch sets the swapper thread's idle transition cost to be the
negative sum of the idle transition costs of all other threads, such
that the sum of idle cost of all threads (swapper + non-swapper threads)
will be zero. Since the sum is always zero, also remove the sum of total
columns.

Bug: 393391273
Change-Id: Id6db241160d0ec003ad031fb6b79039b378c6712
Signed-off-by: Samuel Wu <wusamuel@google.com>