Set VMADDR_FLAG_TO_HOST when doing VM-to-VM VSOCK in QNX
VM-to-VM VSOCK communication in QNX needs to be routed through the
hypervisor. This is enabled by setting the VMADDR_FLAG_TO_HOST flag when
defining the sockaddr_vm struct. Consequently, the communication within
the Android VM needs to be handle the same as if it was running natively
in QNX. Therefore, we make the UnixSocketRaw::Connect call blocking,
just like we did for QNX builds.

Also, a new system property: 'persist.traced.hypervisor` is introduced,
which allows each build target to specify the hypervisor it is running
on in the case it is running virtualized. If the system property is set
to 'qnx', then the UnixSocketRaw::Connect call becomes blocking.

The property can be set in a device makefile as such:
    PRODUCT_PRODUCT_PROPERTIES += ro.traced.hypervisor="qnx"

Test: build traced_relay in Android tree and run Android VM
Bug: 391958400
Change-Id: I24a7457b9a129f32dca3b98ebdd8588d462138cc