Add surface_frame_trace_id and display_trace_id to PipelineReporter
This is a follow-up change for https://crrev.com/c/6208011 which will
allow us to join EventLatency and PipelineReporter events.

The source of truth for the definition of the ChromeFrameReporter proto
message lives in the Perfetto repository (unlike ChromeTrackEvent) so
this change needs to be made across both repositories.

Chrome change: https://crrev.com/c/6226043

Bug: 394060351
Change-Id: I233ea2a8c6a56f518f4e2c87008e6931c405df2e