ui: Make mipmap table empty range a non-error
Currently mipmap tables error out if queried with an empty range.
However, using an empty range isn't inherently a problem, and, in fact,
the mipmap table simply returns an empty list which is exactly what you'd expect it to do.

This CL removes this check, which means we no longer need to guard for
this on the frontend, avoiding some complexity.

Change-Id: I807030db6dd7af4523811ffcc1099a881bf4d8cc