tp: Better select_columns handling in StructuredQuery when using group by
Previous implementation was buggy - it didn't work when the provided select columns didn't have an alias. This approach is the closest to what user would expect when selecting columns when using group by. If the column is a group by column, it can be selected and aliased. If it's a new column added in aggregation, aliasing should not be supported.

Change-Id: I89ecbfe396ddee0e2e534928744c54036404a8bf