Android.bp: generate Android Java SDK related targets.
This CL adds a code to translate 'perfetto_android_jni_library',
'perfetto_android_library' and 'perfetto_android_binary' to the
Android.bp rules with the following logic:

1. Android.bp 'cc_library_shared' rule produces binary with the name
inferred from the rule name, so this CL changes the name for the
targets that produces the JNI libs, to get the expected names.

2. Android.bp 'android_app' rule requires all the apps JNI deps to be
listed in the 'jni_libs' argument, so this CL implements the logic that
filter JNI deps from 'static_libs' and attach them to the 'jni_libs'
for the root 'android_app'.

Tested:
tapas perfetto_java_sdk_app
m
Bug: b/386353531
Change-Id: I8461ff0d518da20e6a6eed5e8da93a3cddc51b28