ui: Added a 'Query Log' tab for debugging slow queries
This tab shows a live view of the most recent 100 queries that have
run, their status, and how long each one took. This is really useful
for diagnosing issues like 'why is this X slow' because it allows us to
see which query is currently running and which what previous queries
were slow.

This tab works in a similar way to the 'Debug SQL Performance' command
except it's live (the list of queries updates in real time), and it's
managed on the client side so it can be used to see the currently
running command which is not possible using the latter method as that
requires running a query itself, which will be blocked by the current
command. It's a catch-22.

Change-Id: I81bb12384c713e67a0a5a325928b428409ef0c96