ui: Fix bug where tracks can get hidden behind sticky headers
When scrolling a track into view (with the F key or otherwise) the
track in question can be hidden behind the sticky group headers.

This CL fixes the issue with the scroll-margin-top css property, which
controls how much margin to leave at the top of the element when
using element.scrollIntoView().

Change-Id: I16a2a4d3f4c1e9ffa6a1001e534e7628e652ee4f