stdlib: Add GPU system state
Add GPU system state, the combination of the GPU frequency and power
state at any given point in time. This is analagous to the CPU system
state.

Make some modifications to the state and derive the actual state of the
hardware instead of directly following what the trace event is
reporting, which is slightly different from the state the hardware is in
from a power context.

Bug: 394608157
Signed-off-by: Samuel Wu <wusamuel@google.com>