ui: Don't allow timeline element to scroll horizontally
If something overflows the bounds of the timeline (usual offenders here
are long track titles), this can cause the timeline to start scrolling
horizontally, which interferes with timeline panning.

This CL simply sets the horizontal overflow to hidden (rather than
visible) in the virtual overlay canvas, which fixes the issue.

Fixes: 394316738
Change-Id: I4311bb84a95b0fafe883de9e629b7025ef439981