Add debug track for frame with max dur blocking call in per-frame metric
For the per-frame metric, apart from pinning the CUJ and blocking calls,
it is useful to pin the frame which contains the blocking call with the
max duration. For the aggregation case 'max_dur_per_frame_ns', added
support to pin the frame during which the blocking call with the max
duration occurred in a separate debug track.

Test: Manual trace and metric check
Bug: 394866952