[ui] Simplify handling id columns in table viewer
Old logic in the table viewer relied on fetching additional columns
(e.g. ts and dur) to display the ID columns and to be able to jump to
them.

This is no longer necessary, so remove the logic which fetches
unnecessary data. Also add tests for handling ID columns for the key
tables (slice, thread_state, sched, process, thread).

Bug: 395565690
Change-Id: Ie768d6df8d93e96bfe401e0f90e79cf54cc8daec