ui: Tweak search iteration behavior
Currently, when iterating through search results, we only iterate
through the results that were present in the visible window as it was
when the search was first started. This is unexpected because the UI
stores some hidden state and doesn't make it obvious to the user.

We also show the total count of search results in the entire trace, rather that just the count of results in this hidden visible window,
which is doubly confusing as it teases more results but doesn't allow
users to access them.

Search needs an entire overhaul, but for now, to address the bug
reports, this CL removes any visible window constraints, so that search
simply searches the entire trace every time, and iteration iterates
through all search results. The only time we consider the visible
window is when choosing which result to start at - we choose the first
event which exists after the start of the visible window after a new
search is made or after changing the search term.

This corresponds more closely to the expected behavior according to
the bug reports.

Fixes: 391704616
Change-Id: Ie7478601688f62803fb8d6a7e64276e2f1cd6a42