tp: relax requirements for emitting clock snapshots
This removes the requirement that clock snapshots must be
emitted before the custom clock is referred to by any `TracePacket`
written by the same `TraceWriter`.

This is accomplished by deferring the processing of packets
that refer to a clock id that is missing a snapshot. We also
enable full sorting mode when such a packet is encountered
to ensure that it will be processed correctly after adding
the missing clock snapshot.

Change-Id: Ia4381a165a520647d8f730e88ccb2d4da12e2bf5