ui: Add proper interface for area selection details tabs
This CL adds a new interface registerAreaSelectionTab() which plugins
can use to add area selection details tabs under the area selection
panel which appears in the current selection tab when an area is
currently selected.

This interface replaces the old registerAggregator() and
registerDetailsPanel(), and an adapter is provided to adapt old
aggregations to the new tab interface.

Compared to the old aggregation interface, this new interface is a lot
more flexible, simply allowing any arbitrary mithril vnodes to be
rendered in the tab's content area, which permits more complex
aggregation views such as flamegraphs.

As a result, the various flamegraphs have been moved out of the core
and into their relevant plugins.

Change-Id: I61bd51b0bba0b3649bfbdd47f6dc48ddc2cb0d83