tp: dpu/tracing_mark_write: ignore tid<->tgid associations during parsing
These events originate from days when the UI track grouping was coupled
to process/thread associations, so some drivers ended up hardcoding
bogus tgids in the event to "group" them in a UI track. Currently we're
parsing those events following the normal B/E/I/C atrace assumptions,
where we trust that the tid of the emitting thread is in the tgid of the
event's body. However since this doesn't hold, we end up with the wrong
thread structure.

This patch selectively overrides the tgid to ignore the incorrect
associations via the existing swapper workarounds.

Bug: 395779936
Change-Id: Ia20dd5e9b2290c467497db64859d15b7df2c95eb