ui: Remove default details panel for DatasetSliceTrack
The default details panel for DatasetSliceTrack assumes slices come from
the slice table, which is not always the case, and degrades the
experience of using the DatasetSliceTrack for arbitrary tables.

This CL removes the default details panel, and pushes the onus onto
the various implementations to select the correct panel for their
use case.

Change-Id: If075944b3bed22b03a3c291817f703652440704b