ui: Add basic syntax highlighting support to the query page
This approach defines a very simple grammar using lezer (see
https://lezer.codemirror.net/docs/guide/), which performs just enough
token-level parsing to get SQL syntax highlighting working.

It is rather incomplete, but supports some of the most common SQLite
keywords, functions, and includes support for some PerfettoSQL
extensions such as macros, and the 'perfetto' keyword.

Despite this being an incomplete solution, the stakes are pretty low if
something goes wrong. Lezer features reboust error recovery, so there's
not much chance of a crash, and the worst that can happen is some
missing syntax highlighting.

Seeing as this is strictly better than what we had before, we should
land this even if we're ultimately going to choose a different solution
in the future.

Bug: 397666766
Change-Id: I4a8f7d748752826c2602acb7e53c4f67c3c4fa0d